+++ src/components/Select/Select.tsx
@@ -27,11 +27,8 @@ const Group = SelectPrimitive.Group;
 
 const Value = SelectPrimitive.Value;
 
-type TriggerProps = React.ComponentPropsWithoutRef<typeof SelectPrimitive.Trigger>;
-const Trigger = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.Trigger>,
-  TriggerProps
->(({ className, children, ...props }, ref) => {
+type TriggerProps = React.ComponentPropsWithRef<typeof SelectPrimitive.Trigger>;
+function Trigger({ ref, className, children, ...props }: TriggerProps) {
   return (
     <SelectPrimitive.Trigger
       ref={ref}
@@ -48,108 +45,129 @@ const Trigger = React.forwardRef<
       </SelectPrimitive.Icon>
     </SelectPrimitive.Trigger>
   );
-});
+}
 Trigger.displayName = SelectPrimitive.Trigger.displayName;
 
-const ScrollUpButton = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.ScrollUpButton>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.ScrollUpButton>
->(({ className, ...props }, ref) => (
-  <SelectPrimitive.ScrollUpButton
-    ref={ref}
-    className={cn('flex cursor-default items-center justify-center py-1', className)}
-    {...props}
-  >
-    <UpOutlined className="size-4" />
-  </SelectPrimitive.ScrollUpButton>
-));
-ScrollUpButton.displayName = SelectPrimitive.ScrollUpButton.displayName;
-
-const ScrollDownButton = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.ScrollDownButton>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.ScrollDownButton>
->(({ className, ...props }, ref) => (
-  <SelectPrimitive.ScrollDownButton
-    ref={ref}
-    className={cn('flex cursor-default items-center justify-center py-1', className)}
-    {...props}
-  >
-    <DownOutlined className="size-4" />
-  </SelectPrimitive.ScrollDownButton>
-));
-ScrollDownButton.displayName = SelectPrimitive.ScrollDownButton.displayName;
-
-const Content = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.Content>
->(({ className, children, position = 'popper', ...props }, ref) => (
-  <SelectPrimitive.Portal>
-    <SelectPrimitive.Content
+function ScrollUpButton({
+  ref,
+  className,
+  ...props
+}: React.ComponentPropsWithRef<typeof SelectPrimitive.ScrollUpButton>) {
+  return (
+    <SelectPrimitive.ScrollUpButton
       ref={ref}
-      className={cn(
-        'relative z-50 max-h-96 min-w-[8rem] overflow-hidden rounded-md border border-slate-200 bg-white text-slate-950 shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2',
-        position === 'popper' &&
-          'data-[side=bottom]:translate-y-1 data-[side=left]:-translate-x-1 data-[side=right]:translate-x-1 data-[side=top]:-translate-y-1',
-        className,
-      )}
-      position={position}
+      className={cn('flex cursor-default items-center justify-center py-1', className)}
       {...props}
     >
-      <ScrollUpButton />
-      <SelectPrimitive.Viewport
+      <UpOutlined className="size-4" />
+    </SelectPrimitive.ScrollUpButton>
+  );
+}
+ScrollUpButton.displayName = SelectPrimitive.ScrollUpButton.displayName;
+
+function ScrollDownButton({
+  ref,
+  className,
+  ...props
+}: React.ComponentPropsWithRef<typeof SelectPrimitive.ScrollDownButton>) {
+  return (
+    <SelectPrimitive.ScrollDownButton
+      ref={ref}
+      className={cn('flex cursor-default items-center justify-center py-1', className)}
+      {...props}
+    >
+      <DownOutlined className="size-4" />
+    </SelectPrimitive.ScrollDownButton>
+  );
+}
+ScrollDownButton.displayName = SelectPrimitive.ScrollDownButton.displayName;
+
+function Content({
+  ref,
+  className,
+  children,
+  position = 'popper',
+  ...props
+}: React.ComponentPropsWithRef<typeof SelectPrimitive.Content>) {
+  return (
+    <SelectPrimitive.Portal>
+      <SelectPrimitive.Content
+        ref={ref}
         className={cn(
-          'p-1',
+          'relative z-50 max-h-96 min-w-[8rem] overflow-hidden rounded-md border border-slate-200 bg-white text-slate-950 shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2',
           position === 'popper' &&
-            'h-[var(--radix-select-trigger-height)] w-full min-w-[var(--radix-select-trigger-width)]',
+            'data-[side=bottom]:translate-y-1 data-[side=left]:-translate-x-1 data-[side=right]:translate-x-1 data-[side=top]:-translate-y-1',
+          className,
         )}
+        position={position}
+        {...props}
       >
-        {children}
-      </SelectPrimitive.Viewport>
-      <ScrollDownButton />
-    </SelectPrimitive.Content>
-  </SelectPrimitive.Portal>
-));
+        <ScrollUpButton />
+        <SelectPrimitive.Viewport
+          className={cn(
+            'p-1',
+            position === 'popper' &&
+              'h-[var(--radix-select-trigger-height)] w-full min-w-[var(--radix-select-trigger-width)]',
+          )}
+        >
+          {children}
+        </SelectPrimitive.Viewport>
+        <ScrollDownButton />
+      </SelectPrimitive.Content>
+    </SelectPrimitive.Portal>
+  );
+}
 Content.displayName = SelectPrimitive.Content.displayName;
 
-const Label = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.Label>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.Label>
->(({ className, ...props }, ref) => (
-  <SelectPrimitive.Label
-    ref={ref}
-    className={cn('px-2 py-1.5 text-[14px] font-semibold', className)}
-    {...props}
-  />
-));
+function Label({
+  ref,
+  className,
+  ...props
+}: React.ComponentPropsWithRef<typeof SelectPrimitive.Label>) {
+  return (
+    <SelectPrimitive.Label
+      ref={ref}
+      className={cn('px-2 py-1.5 text-[14px] font-semibold', className)}
+      {...props}
+    />
+  );
+}
 Label.displayName = SelectPrimitive.Label.displayName;
 
-const Item = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.Item>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.Item>
->(({ className, children, ...props }, ref) => (
-  <SelectPrimitive.Item
-    ref={ref}
-    className={cn(
-      'relative flex w-full cursor-default select-none items-center rounded-sm px-2 py-1.5 text-[14px] outline-none focus:bg-slate-100 focus:text-slate-900 data-[disabled]:pointer-events-none data-[disabled]:opacity-50',
-      className,
-    )}
-    {...props}
-  >
-    <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
-  </SelectPrimitive.Item>
-));
+function Item({
+  ref,
+  className,
+  children,
+  ...props
+}: React.ComponentPropsWithRef<typeof SelectPrimitive.Item>) {
+  return (
+    <SelectPrimitive.Item
+      ref={ref}
+      className={cn(
+        'relative flex w-full cursor-default select-none items-center rounded-sm px-2 py-1.5 text-[14px] outline-none focus:bg-slate-100 focus:text-slate-900 data-[disabled]:pointer-events-none data-[disabled]:opacity-50',
+        className,
+      )}
+      {...props}
+    >
+      <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
+    </SelectPrimitive.Item>
+  );
+}
 Item.displayName = SelectPrimitive.Item.displayName;
 
-const Separator = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.Separator>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.Separator>
->(({ className, ...props }, ref) => (
-  <SelectPrimitive.Separator
-    ref={ref}
-    className={cn('-mx-1 my-1 h-px bg-slate-100', className)}
-    {...props}
-  />
-));
+function Separator({
+  ref,
+  className,
+  ...props
+}: React.ComponentPropsWithRef<typeof SelectPrimitive.Separator>) {
+  return (
+    <SelectPrimitive.Separator
+      ref={ref}
+      className={cn('-mx-1 my-1 h-px bg-slate-100', className)}
+      {...props}
+    />
+  );
+}
 Separator.displayName = SelectPrimitive.Separator.displayName;
 
 const Select = {