+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Changed SQL editor auto-complete to use prefix matching and require at least
   2 characters before showing suggestions.
+- bump React to version 19
 
 ## 2026-03-31 - 0.23.3