+++ CHANGES.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
+## 2026-04-16 - 0.24.0
+
 - Changed SQL editor auto-complete to use prefix matching and require at least
   2 characters before showing suggestions.
-- bump React to version 19
+- Bump React to version 19
 
 ## 2026-03-31 - 0.23.3